05.03 dev DEV-WBENCH6

+++ Workbench/src/main/java/workbench/graphics/scene/renderer/WBenchOpenGLRenderer.java
@@ -33,6 +33,7 @@ import workbench.WBench;
 import workbench.graphics.scene.nodes.templates.*;
 import workbench.graphics.scene.nodes.*;
 import workbench.graphics.scene.ui.EditorInterface;
+import workbench.graphics.scene.ui.LoadingInterface;
 import workbench.graphics.scene.ui.ProjectInitInterface;
 import workbench.graphics.scene.world.WBenchWorld;
 import workbench.project.Project;
@@ -49,6 +50,7 @@ public class WBenchOpenGLRenderer extends OpenGLRenderer implements IDearUIImp, 
 
     private static DearUIInterface editorInterface;
     private static DearUIInterface projectInterface;
+    private static DearUIInterface loadingInterface;
 
     protected Map<NodeID, IRenderNode> conveyorNodes;
     protected IndirectBufferProgram sceneIndirectBufferProgram;
@@ -62,6 +64,7 @@ public class WBenchOpenGLRenderer extends OpenGLRenderer implements IDearUIImp, 
 
         WBenchOpenGLRenderer.editorInterface = new EditorInterface(WBench.get().getProjectManager());
         WBenchOpenGLRenderer.projectInterface = new ProjectInitInterface();
+        WBenchOpenGLRenderer.loadingInterface = new LoadingInterface();
 
         this.sceneIndirectBufferProgram = new IndirectBufferProgram(DefaultAttributePointers.ATTR_POSITIONS, DefaultAttributePointers.ATTR_NORMALS, DefaultAttributePointers.ATTR_TEXTURE_COORDINATES, DefaultAttributePointers.ATTR_TANGENTS, DefaultAttributePointers.ATTR_BI_TANGENTS, DefaultAttributePointers.ATTR_BONES_INDEXES, DefaultAttributePointers.ATTR_BONES_WEIGHTS);
         this.screenModel = null;
@@ -213,7 +216,9 @@ public class WBenchOpenGLRenderer extends OpenGLRenderer implements IDearUIImp, 
     public void onClosingProject(WBenchResourceManager resourceManager, @NotNull Project project) {
         this.destroySceneIndirectRenderBuffer();
 
-        this.getWorld().getEnvironment().destroyEnvironment();
+        if (this.getWorld().getEnvironment() != null) {
+            this.getWorld().getEnvironment().destroyEnvironment();
+        }
         this.getConveyorNodes().values().stream().filter(e -> !(e instanceof IUIRenderNode)).forEach(IRenderNode::destroyResources);
         this.getSceneCulling().destroyResources();
         this.removeNodes();
@@ -299,5 +304,9 @@ public class WBenchOpenGLRenderer extends OpenGLRenderer implements IDearUIImp, 
 
     public static DearUIInterface getProjectInterface() {
         return WBenchOpenGLRenderer.projectInterface;
+    }
+
+    public static DearUIInterface getLoadingInterface() {
+        return WBenchOpenGLRenderer.loadingInterface;
     }
 }
